LayoutTests/imported/w3c/web-platform-tests/css/css-properties-values-api/animation/custom-property-transition-transform-function-matrix-expected.txt


FAIL A custom property of type <transform-function> can yield a CSS Transition between different function types assert_equals: Element has the expected animated value expected "matrix3d(2.5, 0, 0, 0, 0, 2.5, 0, 0, 0, 0, 1, 0, 0, 0, 50, 1)" but got "scale(4)"
